Move expressions into conditions #66
